moving the pdp feed helper
